% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{sparseLRMatrix-class}
\alias{sparseLRMatrix-class}
\title{Sparse plus low rank matrix}
\value{

}
\description{
Eventually this class will subclass \code{Matrix} objects,
but for now this is a basic implementation that essentially
only supports singular value decomposition.
}
\details{
To learn more about S4 classes, please see
\url{https://adv-r.hadley.nz/s4.html}.
}
\section{Slots}{

\describe{
\item{\code{sparse}}{sparseMatrix.}

\item{\code{U}}{Matrix.}

\item{\code{V}}{Matrix.}
}}

