% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseLRMatrix-package.R
\docType{package}
\name{sparseLRMatrix-package}
\alias{sparseLRMatrix}
\alias{sparseLRMatrix-package}
\title{sparseLRMatrix: Represent and Use Sparse + Low Rank Matrices}
\description{
Provides an S4 class for representing and
    interacting with sparse plus rank matrices. At the moment the
    implementation is quite spare, but the plan is eventually subclass
    Matrix objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RoheLab/sparseLRMatrix}
  \item Report bugs at \url{https://github.com/RoheLab/sparseLRMatrix/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (\href{https://orcid.org/0000-0002-4985-5160}{ORCID})

}
\keyword{internal}
