% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{sparseLRMatrix}
\alias{sparseLRMatrix}
\title{Title}
\usage{
sparseLRMatrix(sparse, U, V)
}
\arguments{
\item{sparse}{TODO}

\item{U}{TODO}

\item{V}{TODO}
}
\value{

}
\description{
Title
}
\examples{

# TODO
}
